Update Kotlin version to latest - 1.3.21
This version fixes the build, removing the following warnings:

```
$ ./gradlew check
...
Deprecated Gradle features were used in this build, making it
incompatible with Gradle 6.0.
Use '--warning-mode all' to show the individual deprecation warnings.
See
https://docs.gradle.org/5.1.1/userguide/command_line_interface.html#sec:command_line_warnings
...
$ ./gradlew check  --warning-mode all
...
The DefaultSourceDirectorySet constructor has been deprecated. This is
scheduled to be removed in Gradle 6.0. Please use the ObjectFactory
service to create instances of SourceDirectorySet instead.
...
```